Rewrite of purge-old-kernels
This is Bash script. This one purges requested amount of kernels OLDER than the current kernel, so it does not purge the newest kernel in  case system has not been rebooted after its installation. This one does not need superuser privileges, if you run it in simulate mode. This one has help option. Purges all "^linux-.*" packages corresponding the kernel versions.  Uses --auto-remove in apt-get command; maybe it is excessive?